+++ src/test/java/net/sf/jabref/exporter/BibDatabaseWriterTest.java
@@ -2,17 +2,20 @@ package net.sf.jabref.exporter;
 
 import com.google.common.base.Charsets;
 import net.sf.jabref.*;
+import net.sf.jabref.bibtex.EntryTypes;
+import net.sf.jabref.groups.GroupTreeNode;
+import net.sf.jabref.groups.structure.*;
 import net.sf.jabref.importer.ImportFormatReader;
 import net.sf.jabref.importer.ParserResult;
 import net.sf.jabref.importer.fileformat.BibtexParser;
+import net.sf.jabref.logic.labelPattern.DatabaseLabelPattern;
 import net.sf.jabref.logic.util.io.FileUtil;
 import net.sf.jabref.model.database.BibDatabase;
 import net.sf.jabref.model.database.BibDatabaseMode;
-import net.sf.jabref.model.entry.BibEntry;
-import net.sf.jabref.model.entry.BibtexEntryTypes;
-import net.sf.jabref.model.entry.EntryType;
+import net.sf.jabref.model.entry.*;
 import org.junit.Assert;
 import org.junit.BeforeClass;
+import org.junit.Ignore;
 import org.junit.Test;
 import org.mockito.internal.util.io.IOUtil;
 import sun.misc.IOUtils;
@@ -86,7 +89,7 @@ public class BibDatabaseWriterTest {
         databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), new SavePreferences(), false,
                 false);
 
-        Assert.assertEquals(Globals.NEWLINE + "@PREAMBLE{Test preamble}" + Globals.NEWLINE, stringWriter.toString());
+        Assert.assertEquals(Globals.NEWLINE + "@Preamble{Test preamble}" + Globals.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -102,7 +105,7 @@ public class BibDatabaseWriterTest {
         databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), preferences, false, false);
 
         Assert.assertEquals("% Encoding: US-ASCII" + Globals.NEWLINE + Globals.NEWLINE +
-                "@PREAMBLE{Test preamble}" + Globals.NEWLINE, stringWriter.toString());
+                "@Preamble{Test preamble}" + Globals.NEWLINE, stringWriter.toString());
     }
 
     @Test
@@ -139,6 +142,171 @@ public class BibDatabaseWriterTest {
 
         Assert.assertEquals("% Encoding: US-ASCII" + Globals.NEWLINE + Globals.NEWLINE +
                 "@Article{," + Globals.NEWLINE + "}" + Globals.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    public void writeEpilogue() throws IOException {
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        BibDatabase database = new BibDatabase();
+        database.setEpilog("Test epilog");
+        BibDatabaseContext context = new BibDatabaseContext(database, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), new SavePreferences(), false,
+                false);
+
+        Assert.assertEquals(Globals.NEWLINE + "Test epilog" + Globals.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    public void writeEpilogueAndEncoding() throws IOException {
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        SavePreferences preferences = new SavePreferences();
+        preferences.setEncoding(Charsets.US_ASCII);
+        BibDatabase database = new BibDatabase();
+        database.setEpilog("Test epilog");
+        BibDatabaseContext context = new BibDatabaseContext(database, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), preferences, false, false);
+
+        Assert.assertEquals("% Encoding: US-ASCII" + Globals.NEWLINE + Globals.NEWLINE +
+                "Test epilog" + Globals.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    public void writeMetadata() throws IOException {
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        MetaData metaData = new MetaData();
+        DatabaseLabelPattern labelPattern = new DatabaseLabelPattern();
+        labelPattern.setDefaultValue("test");
+        metaData.setLabelPattern(labelPattern);
+        BibDatabaseContext context = new BibDatabaseContext(new BibDatabase(), metaData, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), new SavePreferences(), false,
+                false);
+
+        Assert.assertEquals(Globals.NEWLINE + "@Comment{jabref-meta: keypatterndefault:test;}" + Globals.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    public void writeMetadataAndEncoding() throws IOException {
+        SavePreferences preferences = new SavePreferences();
+        preferences.setEncoding(Charsets.US_ASCII);
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        MetaData metaData = new MetaData();
+        DatabaseLabelPattern labelPattern = new DatabaseLabelPattern();
+        labelPattern.setDefaultValue("test");
+        metaData.setLabelPattern(labelPattern);
+        BibDatabaseContext context = new BibDatabaseContext(new BibDatabase(), metaData, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), preferences, false, false);
+
+        Assert.assertEquals("% Encoding: US-ASCII" + Globals.NEWLINE + Globals.NEWLINE +
+                "@Comment{jabref-meta: keypatterndefault:test;}" + Globals.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    public void writeGroups() throws IOException {
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        MetaData metaData = new MetaData();
+        GroupTreeNode groupRoot = new GroupTreeNode(new AllEntriesGroup());
+        groupRoot.add(new GroupTreeNode(new ExplicitGroup("test", GroupHierarchyType.INCLUDING)));
+        metaData.setGroups(groupRoot);
+        BibDatabaseContext context = new BibDatabaseContext(new BibDatabase(), metaData, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), new SavePreferences(), false,
+                false);
+
+        // @formatter:off
+        Assert.assertEquals(Globals.NEWLINE +
+                "@Comment{jabref-meta: groupsversion:3;}" + Globals.NEWLINE
+                + Globals.NEWLINE
+                + "@Comment{jabref-meta: groupstree:" + Globals.NEWLINE
+                + "0 AllEntriesGroup:;" + Globals.NEWLINE
+                + "1 ExplicitGroup:test\\;2\\;;" + Globals.NEWLINE
+                + "}" + Globals.NEWLINE, stringWriter.toString());
+        // @formatter:on
+    }
+
+    @Test
+    public void writeGroupsAndEncoding() throws IOException {
+        SavePreferences preferences = new SavePreferences();
+        preferences.setEncoding(Charsets.US_ASCII);
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        MetaData metaData = new MetaData();
+        GroupTreeNode groupRoot = new GroupTreeNode(new AllEntriesGroup());
+        groupRoot.add(new GroupTreeNode(new ExplicitGroup("test", GroupHierarchyType.INCLUDING)));
+        metaData.setGroups(groupRoot);
+        BibDatabaseContext context = new BibDatabaseContext(new BibDatabase(), metaData, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), preferences, false, false);
+
+        // @formatter:off
+        Assert.assertEquals(
+                "% Encoding: US-ASCII" + Globals.NEWLINE +
+                Globals.NEWLINE +
+                "@Comment{jabref-meta: groupsversion:3;}" + Globals.NEWLINE
+                + Globals.NEWLINE
+                + "@Comment{jabref-meta: groupstree:" + Globals.NEWLINE
+                + "0 AllEntriesGroup:;" + Globals.NEWLINE
+                + "1 ExplicitGroup:test\\;2\\;;" + Globals.NEWLINE
+                + "}" + Globals.NEWLINE, stringWriter.toString());
+        // @formatter:on
+    }
+
+    @Test
+    public void writeString() throws IOException {
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        BibDatabase database = new BibDatabase();
+        database.addString(new BibtexString("id", "name", "content"));
+        BibDatabaseContext context = new BibDatabaseContext(database, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), new SavePreferences(), false,
+                false);
+
+        Assert.assertEquals(Globals.NEWLINE + "@String{name = {content}}" + Globals.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    public void writeStringAndEncoding() throws IOException {
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        SavePreferences preferences = new SavePreferences();
+        preferences.setEncoding(Charsets.US_ASCII);
+        BibDatabase database = new BibDatabase();
+        database.addString(new BibtexString("id", "name", "content"));
+        BibDatabaseContext context = new BibDatabaseContext(database, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.emptyList(), preferences, false, false);
+
+        Assert.assertEquals("% Encoding: US-ASCII" + Globals.NEWLINE + Globals.NEWLINE +
+                "@String{name = {content}}" + Globals.NEWLINE, stringWriter.toString());
+    }
+
+    @Test
+    public void writeEntryWithCustomizedTypeAlsoWritesTypeDeclaration() throws IOException {
+        BibDatabaseWriter databaseWriter = new BibDatabaseWriter();
+        EntryTypes.addOrModifyCustomEntryType(new CustomEntryType("customizedType", "required", "optional"));
+        BibEntry entry = new BibEntry();
+        entry.setType("customizedType");
+        BibDatabase database = new BibDatabase();
+        database.insertEntry(entry);
+        BibDatabaseContext context = new BibDatabaseContext(database, new Defaults(BibDatabaseMode.BIBTEX));
+
+        StringWriter stringWriter = new StringWriter();
+        databaseWriter.writePartOfDatabase(stringWriter, context, Collections.singletonList(entry),
+                new SavePreferences(), false, false);
+
+        Assert.assertEquals(Globals.NEWLINE +
+                        "@Customizedtype{," + Globals.NEWLINE + "}" + Globals.NEWLINE + Globals.NEWLINE
+                        + "@Comment{jabref-entrytype: Customizedtype: req[required] opt[optional]}" + Globals.NEWLINE,
+                stringWriter.toString());
     }
 
     @Test